Highlight expressions based on value equality
I.e., when hovering, highlight all expressions in the
same static frame (i.e. the same function body) with the
same value. This also results in functions not being highlighted
as their equality can't be determined. This is a bit of an
experiment.

This commit also moves Value to it's own file and updates the
Dict import in Reader.elm because I missed it 2 commits ago.